+++ lib/loader/platform.ts
@@ -1,0 +1,3 @@
+import { platform } from 'node:os';
+
+export const __isWindows = platform() === 'win32';